Refactor and restructure in prep for tower defense
