+++ op-challenger/fault/chain/loader.go
@@ -1,4 +1,4 @@
-package fault
+package chain
 
 import (
 	"context"
