+++ op-challenger/fault/responder/responder.go
@@ -1,4 +1,4 @@
-package fault
+package responder
 
 import (
 	"context"
