+++ op-challenger/fault/responder/responder_test.go
@@ -1,4 +1,4 @@
-package fault
+package responder
 
 import (
 	"context"
