+++ op-challenger/fault/types/common.go
@@ -1,7 +1,6 @@
 package types
 
 import (
-	"context"
 	"errors"
 	"math/big"
 
@@ -12,13 +11,27 @@ var (
 	ErrGameDepthReached = errors.New("game depth reached")
 )
 
-type GameStatus uint8
-
 const (
 	GameStatusInProgress GameStatus = iota
 	GameStatusChallengerWon
 	GameStatusDefenderWon
 )
+
+type GameStatus uint8
+
+// GameStatusString returns the current game status as a string.
+func GameStatusString(status GameStatus) string {
+	switch status {
+	case GameStatusInProgress:
+		return "In Progress"
+	case GameStatusChallengerWon:
+		return "Challenger Won"
+	case GameStatusDefenderWon:
+		return "Defender Won"
+	default:
+		return "Unknown"
+	}
+}
 
 // PreimageOracleData encapsulates the preimage oracle data
 // to load into the onchain oracle.
@@ -60,28 +73,6 @@ type StepCallData struct {
 	IsAttack   bool
 	StateData  []byte
 	Proof      []byte
-}
-
-// OracleUpdater is a generic interface for updating oracles.
-type OracleUpdater interface {
-	// UpdateOracle updates the oracle with the given data.
-	UpdateOracle(ctx context.Context, data *PreimageOracleData) error
-}
-
-// TraceProvider is a generic way to get a claim value at a specific step in the trace.
-type TraceProvider interface {
-	// Get returns the claim value at the requested index.
-	// Get(i) = Keccak256(GetPreimage(i))
-	Get(ctx context.Context, i uint64) (common.Hash, error)
-
-	// GetStepData returns the data required to execute the step at the specified trace index.
-	// This includes the pre-state of the step (not hashed), the proof data required during step execution
-	// and any pre-image data that needs to be loaded into the oracle prior to execution (may be nil)
-	// The prestate returned from GetStepData for trace 10 should be the pre-image of the claim from trace 9
-	GetStepData(ctx context.Context, i uint64) (prestate []byte, proofData []byte, preimageData *PreimageOracleData, err error)
-
-	// AbsolutePreState is the pre-image value of the trace that transitions to the trace value at index 0
-	AbsolutePreState(ctx context.Context) (preimage []byte, err error)
 }
 
 // ClaimData is the core of a claim. It must be unique inside a specific game.
