+++ op-challenger/fault/types/common_test.go
