+++ op-challenger/fault/utils/checks_test.go
@@ -1,4 +1,4 @@
-package fault
+package utils
 
 import (
 	"context"
